[ftinspect] Improve Waterfall mode.
* src/ftinspect/rendering/glyphcontinuous.cpp,
  src/ftinspect/rendering/glyphcontinuous.hpp:
  Don't paint the red square (placeholder for zero x-advance glyphs) when in
  Waterfall mode.
  Use the `StringRenderer`'s line begin callback to render a font point size
  prefix when in Waterfall mode. `sizeIndicatorOffset_` shifts the whole
  string right.

* src/ftinspect/engine/stringrenderer.cpp: Fix `pointSize` return type.
  Add `isWaterfall` func.
  Add support to line begin hook callback.

* src/ftinspect/engine/engine.hpp: `pointSize` func should return `double`.

+++ src/ftinspect/engine/engine.hpp
@@ -108,7 +108,7 @@ public:
   FT_Library ftLibrary() const { return library_; }
   FTC_Manager cacheManager() { return cacheManager_; }
   int dpi() { return dpi_; }
-  int pointSize() { return pointSize_; }
+  double pointSize() { return pointSize_; }
   int currentFontType() const { return fontType_; }
   const QString& currentFamilyName() { return curFamilyName_; }
   const QString& currentStyleName() { return curStyleName_; }
